requirements.txt
playwright>=1.40
Pillow>=10.0
openpyxl>=3.1
pytest>=7.4
